Slight updates to MinGW.
Now you must set the environment variables MINGW_TRIPLE and
MINGW_LOCATION before calling CMake. This is untested as I don't have a
MinGW Audacious, and is really unnecessary since Audacious now includes
this plugin; but if somebody is looking to compile it for whatever
reason, this might give you a fighting chance.